+++ jni/share/spellingtrie.cpp
@@ -554,9 +554,9 @@ namespace ime_pinyin {
     bool SpellingTrie::save_spl_trie ( FILE *fp ) {
         if ( NULL == fp || NULL == spelling_buf_ )
         { return false; }
-        if ( fwrite ( &spelling_size_, sizeof ( size_t ), 1, fp ) != 1 )
+        if ( fwrite ( &spelling_size_, sizeof ( int ), 1, fp ) != 1 )
         { return false; }
-        if ( fwrite ( &spelling_num_, sizeof ( size_t ), 1, fp ) != 1 )
+        if ( fwrite ( &spelling_num_, sizeof ( int ), 1, fp ) != 1 )
         { return false; }
         if ( fwrite ( &score_amplifier_, sizeof ( float ), 1, fp ) != 1 )
         { return false; }
